+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ------------------
 
 - drop support for numpy 1.25. [#285]
+- require asdf 3.3 and drop support for astropy 5.2 [#297]
 
 0.8.0 (2025-06-11)
 ------------------